client-service/requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
flake8==6.1.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mistune==3.0.1
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
